[IMP] portal_rating: allow to tweak popup composer button display
PURPOSE

Change the display of the button to open the composer from btn-primary to
btn-link.

SPECIFICATIONS

In order to ease integration of popup composer rating widget, let us allow
to customize classes given to the button triggering it. In slides we want
specifically it to be a link, not a button.

LINKS

Task ID-2241513
PR #55698