[IMP] account: tax report: add the possibility to use a different country than the company's for the tax report
Some companies have their office in a country, and sell and pay taxes in another one. Before this commit, they couldn't use Odoo's tax report, as the country set on the res.company than the one to be used with their taxes. It is now possible to do that, using a config parameter.

closes odoo/odoo#52734

X-original-commit: 6fb601786a8b178d18156720d79dfd6c9a447e76
Related: odoo/enterprise#11039
Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>
Signed-off-by: oco-odoo <oco-odoo@users.noreply.github.com>