[FIX] web_editor: make a test more robust
Don't know how this test worked before but it will stop working with
next updates. This just make it more robust using the same trick as
other tests: wait for the colorpicker to be opened before actually
using it.

Part of https://github.com/odoo/odoo/pull/52211
Related to task-2197038