[FIX] base: restore _clean_website on res.partner
Commit 69c1727a82 improved the UrlWidget s.t. it automatically
prefixes the url with 'http://' if no protocol is specified (except
for own website paths).

In addition, it removed a logic specific to model res.partner that
automatically prefixes the saved website values with 'http://'.

However, the second part should not have been done: those urls can
be displayed without UrlWidget (e.g. on the website), and in this
case, the protocol wouldn't be set, meaning that it would behave
like a relative link.

closes odoo/odoo#52113

X-original-commit: d51d091448009f6f617ad69e4f129222cac73e98
Signed-off-by: Jérémy Kersten (jke) <jke@openerp.com>