[FIX] web_tour: tooltip bubble info glitch and moves to the left
When you hover on the bubble tooltip infobox then sometimes its glitch,
and moves left. once you leave the mouse over from the tooltip info then
it will be repositioned as before.

So with this commit, we have resolved this issue by giving the condition
on the repositioning of every time the _updatePosition methods execute.
If the tooltip infobox is open then we are not repositioning the box's position.
otherwise, it will be repositioned.

closes odoo/odoo#55162

Task: 2263600
X-original-commit: 669030550f7dd21c20f743fbfc444ef5c0f1fa08
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>