[FIX] sale_timesheet: Display of SOL field on task only if timesheet exist.
Currently, the SOL field is visible on task form view even if there are
no timesheets linked to the task.

So in this commit the SOL field should not be visible on task form view
if there are no timesheets linked to the task. The SOL field is only
visible incase when there is multiple soline is linked with timesheet
so to make that field visibile/invisible there is has_multi_sol
compute field. so we will only display the field if timesheet_id is
exist.

closes odoo/odoo#52218

Taskid: 2266104
Closes: #52218
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>