[IMP] mail: add possibility to choose a source / request partner on activity
PURPOSE

Allow to know if an activity is linked to a customer or a requesting partner.

SPECIFCIATIONS

Add a partner_id field on mail.activity model. It will be used notably in
website slides to know who triggered an "request access" activity. It
can also be used in various business flows when assigned user is different
from the "source" or "requesting" partner.

LINKS

Task ID : 2127364
Community PR odoo/odoo#40708
Upgrade PR odoo/upgrade#1101